CP-50091: Moved inventory.py from scripts/examples/python/ to python3/packages/inventory.py - Fixed bare-except, unspecified-encoding  and indentation issue
